Task-11: Implement an endpoint of type POST
There is not only a GET request, but there are also the following types: GET, POST, PUT, PATCH, DELETE and others.

We will look at the most used ones. Let's add a POST request in which we will pass the request body, in which we will place:

-user id

-user name

-User API_KEY

In response, the user should receive a picture from the rover.